% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanerWetlands.R
\docType{package}
\name{cleanerWetlands}
\alias{cleanerWetlands}
\title{\code{cleanerWetlands} package}
\description{
This package allows us to clean up data provided by regional councils before entering it into the MWLR databaese.
}
