% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_HBRC_front_matter.R
\name{read_HBRC_front_matter}
\alias{read_HBRC_front_matter}
\title{Read in wetland-level metadata from HBRC files}
\usage{
read_HBRC_front_matter(
  filepath,
  frontsheet = 1,
  returnConditionDetailed = FALSE
)
}
\arguments{
\item{filepath}{Path to the excel file, including file name and extension}

\item{frontsheet}{Which sheet of the excel file contains the main metadata. Defaults to 1.}

\item{returnConditionDetailed}{Whether to return a simple metadata table (defaults to this; ie TRUE), or a more detailed table of the elements of condition, in which case a list is returned.}
}
\value{
Simple metadatatable (unless a detailed version is required)
}
\description{
Read in wetland-level metadata from HBRC files
}
