% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_HBRC_veg_plots.R
\name{read_HBRC_veg_plots}
\alias{read_HBRC_veg_plots}
\title{Read in wetland plot data from HBRC files}
\usage{
read_HBRC_veg_plots(filepath)
}
\arguments{
\item{filepath}{Path to the excel file, including file name and extension}
}
\value{
List with three parts "vegPlotData" (veg recce data), "soilSamples" (id of soil samples collected on each plot), "vegSamples" (id of plant foliage samples collected on each plot)
}
\description{
There is a quirk with percentage data: if it has been formatted as a percentage in excel, read_excel(), which is used under the hood, doesn't pick up the formatting in complex sheets. Thus it either comes in as proportion or percentage with no indication. This function checks (based on summed total cover) and adds a warning where values are assumed to be proportion, and have therefore been multiplied by 100 to get to percentage.
}
